code split of search bar component
