+++ app/client/src/ee/pages/Homepage/DesktopEntitySearchField.tsx
@@ -1,0 +1,3 @@
+export * from "ce/pages/Homepage/DesktopEntitySearchField";
+import { default as CE_DesktopEntitySearchField } from "ce/pages/Homepage/DesktopEntitySearchField";
+export default CE_DesktopEntitySearchField;
