+++ app/client/src/ee/pages/Homepage/MobileEntitySearchField.tsx
@@ -1,0 +1,3 @@
+export * from "ce/pages/Homepage/MobileEntitySearchField";
+import { default as CE_MobileEntitySearchField } from "ce/pages/Homepage/MobileEntitySearchField";
+export default CE_MobileEntitySearchField;
